backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.2.1

langchain==0.2.11
langchain-core==0.2.26
langchain-community==0.2.10
langchain-google-genai==1.0.8
langchain-chroma==0.1.4

chromadb==0.5.3

google-generativeai==0.7.2

pypdf==4.2.0
python-dotenv==1.0.1

pymongo==4.7.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0